2014-03-17, 21:33
-worked on autonomous
-minor bugfixes
-cleanup
-auto drive has notes inside for what to do tomorrow, attempt to make it
drive ligitimatly, it was just being given an excessive setpoint to get
there fast enough, we are just going to change P and give the actual
setpoint. need to give rate PID on right a significant advantage.
Instead of putting it in hold, leave it in auto drive (consider doing)